Add extended support for X.509 SANs
With these changes, it is possible to read email, URN and DNS Subject Alternative Names from X.509 certificates. An existing function has been moved and extended to be usable from outside.

A test had to be adjusted, because it now (correctly) parses an email SAN in one of the test certificates.